fix : Updateのデルタタイムをuint64からfloatに変更
feat : 簡易的な昼夜表現実装

feat : ゲーム内の時間とタイムスケールを管理するTimeクラスの実装

+++ 3D_TPS/system/Framework/Game/Game.h
@@ -17,7 +17,7 @@ public:
 	~Game();
 
 	void Init(void);
-	void Update(const uint64_t deltatime);
+	void Update(const float deltatime);
 	void Draw(void);
 	void Uninit(void);
