+++ 3D_TPS/system/Framework/ObjectManager/ObjectManager.h
@@ -52,7 +52,7 @@ public:
 	//void Init(ComponentFactory* _factory);
 	//void Init(ShaderManager* shaderMgr);
 	void Init(EngineContext* context);
-	void Update(const uint64_t deltatime);
+	void Update(const float deltatime);
 	void Draw(void) const;
 	void Uninit(void);
