+++ 3D_TPS/system/GameObject/Skydome.h
@@ -16,7 +16,7 @@ public:
 	~Skydome();
 
 	void Init(void) override;
-	void Update(const uint64_t deltatime) override;
+	void Update(const float deltatime) override;
 	void Draw(void) const override;
 	void Uninit(void) override;
